Port shaping profile
